fix: 优化录像播放高度
perf: 去掉log
