+++ src/app/elements/term/term.component.html
@@ -1,4 +1,4 @@
 <div style='padding:15px 0 15px 15px'>
-  <div #term (mouseenter)="active()"></div>
+  <div #term id="terminal" (mouseenter)="active()"></div>
 </div>
